Migrated project to vite
