+++ package.json
@@ -13,6 +13,8 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "@vitejs/plugin-react-swc": "^3.7.0",
     "assert": "^2.0.0",
     "crypto": "^1.0.1",
     "crypto-browserify": "^3.12.0",
@@ -27,19 +29,19 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.0",
-    "react-scripts": "5.0.1",
-    "react-textarea-autosize": "^8.5.3",
     "stream": "^0.0.2",
     "stream-browserify": "^3.0.0",
     "styled-components": "^5.3.11",
     "util": "^0.12.5",
+    "vite": "^5.3.4",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.2",
     "web-vitals": "^3.3.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview"
   },
   "eslintConfig": {
     "extends": [
