+++ src/App.jsx
@@ -6,10 +6,21 @@ import Login from "./pages/Login";
 import PrivateRoute from "./PrivateRoute";
 import Layout from "./Layout";
 import { useSelector } from "react-redux";
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
+
+const darkTheme = createTheme({
+  palette: {
+    mode: 'dark',
+  },
+});
+
 
 function App() {
     const loggedIn = useSelector((state) => state.loggedIn);
     return (
+        <ThemeProvider theme={darkTheme}>
+        <CssBaseline />
         <BrowserRouter>
             <Layout />
             <Routes>
@@ -18,6 +29,7 @@ function App() {
                 <Route path='/signup' element={<Signup />} />
             </Routes>
         </BrowserRouter>
+        </ThemeProvider>
     );
 }
