+++ src/BoardSelector.jsx
@@ -52,7 +52,6 @@ export const BoardSelector = (props) => {
     };
 
     const updateSelectedChallenge = (e) => {
-        console.log(e.target.childNodes[0].data);
         dispatch({
             type: "setSelectedChallenge",
             payload: e.target.childNodes[0].data,
