+++ src/EditResultDialog.jsx
@@ -8,31 +8,67 @@ import {
     Select,
     DialogActions,
 } from "@mui/material";
+import { useEffect } from "react";
 import { TimePicker, LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 
 import { useState } from "react";
-import { getDatabase, ref, set, push, onValue } from "firebase/database";
+import {
+    getDatabase,
+    ref,
+    set,
+    push,
+    onValue,
+    update,
+} from "firebase/database";
 import { useDispatch, useSelector } from "react-redux";
-import TextareaAutosize from "react-textarea-autosize";
+import dayjs from "dayjs";
 
-const NewResultDialog = (props) => {
+const EditResultDialog = (props) => {
     const [challenge, setChallenge] = useState();
     const [type, setType] = useState();
     const [team, setTeam] = useState();
     const [score, setScore] = useState();
-    const [notes, setNotes] = useState("");
-    const [invalid, setInvalid] = useState(false);
-    const dispatch = useDispatch();
     const db = getDatabase();
-    const challengeType = useSelector((state) => state.newResultChallengeType);
+    const challengeMap = useSelector((state) => state.challengeMap);
+    const selectedChallenge = useSelector((state) => state.selectedChallenge);
+    const currentChallenge = useSelector((state) => state.currentChallenge);
+    const [challengeType, setChallengeType] = useState();
+    const [teamId, setTeamId] = useState();
+    const { open, onClose, editTeam, teamType } = props;
 
-    function writeTeamResultData(e) {
-        dispatch({
-            type: "setSelectedChallenge",
-            payload: challenge.replace(/_/g, " "),
+    useEffect(() => {
+        console.log(props.teamType);
+        setTeam(editTeam[0]);
+        setScore(editTeam[1]);
+        setTeamId(editTeam[2]);
+        if (selectedChallenge != null) {
+            setChallenge(selectedChallenge);
+        } else if (currentChallenge != null) {
+            setChallenge(currentChallenge);
+        }
+    }, []);
+
+    useEffect(() => {
+        let chalType = null;
+        challengeMap.forEach((chal) => {
+            if (chal[0] === challenge) {
+                chalType = chal[1];
+            }
         });
+        if (chalType === null) {
+            setChallengeType(challengeMap[selectedChallenge]);
+        } else {
+            setChallengeType(chalType);
+        }
+        if (chalType === "Time") {
+            secondsToMinutes(score);
+        }
+    }, [challenge]);
+
+    function updateTeamResults(e) {
         e.preventDefault();
+        console.log(teamType);
         let newScore = null;
 
         if (challengeType === "Time") {
@@ -42,14 +78,18 @@ const NewResultDialog = (props) => {
         }
         const teamListRef = ref(
             db,
-            "Challenges/" + challenge + "/" + type + "/"
+            "Challenges/" +
+                challenge.replace(/ /g, "_") +
+                "/" +
+                teamType +
+                "/" +
+                teamId +
+                "/"
         );
-        const newTeamRef = push(teamListRef);
-        set(newTeamRef, {
+
+        set(teamListRef, {
             team: team,
             score: Number(newScore),
-            notes: notes,
-            date: Date.now(),
         });
         props.onClose();
     }
@@ -66,76 +106,31 @@ const NewResultDialog = (props) => {
         setTeam(event.target.value);
     };
 
-    const handleChallengeChange = (event) => {
-        setChallenge(event.target.value.replace(/ /g, "_"));
-    };
-
     const handleTypeChange = (event) => {
         setType(event.target.value);
-    };
-
-    const handleNotesChange = (event) => {
-        setNotes(event.target.value);
     };
 
     const minutesToSeconds = (time) => {
         let minutes = time.get("minutes");
         let seconds = time.get("seconds");
-        console.log(minutes * 60 + seconds);
         return Number(minutes) * 60 + Number(seconds);
     };
 
-    const createChallengeItems = () => {
-        const challengeRef = ref(db, "Challenges/");
-        let challengeList = [];
-        onValue(challengeRef, (snapshot) => {
-            const challenges = snapshot.val();
-            for (const key in challenges) {
-                challengeList.push([
-                    key.replace(/_/g, " "),
-                    challenges[key]["type"],
-                ]);
-            }
-        });
-
-        return challengeList.map((challenge) => {
-            return (
-                <MenuItem
-                    key={challenge[0]}
-                    data-type={challenge[1]}
-                    value={challenge[0]}
-                    onClick={(e) =>
-                        dispatch({
-                            type: "setNewResultChallengeType",
-                            payload: e.target.dataset.type,
-                        })
-                    }
-                >
-                    {challenge[0]}
-                </MenuItem>
-            );
-        });
+    const secondsToMinutes = (time) => {
+        let minutes = Math.floor(time / 60);
+        let seconds = time - minutes * 60;
+        if (seconds < 10) {
+            seconds = "0" + seconds;
+        } else if (seconds === 0) {
+            seconds = "00";
+        }
+        setScore(dayjs().minute(minutes).second(Number(seconds)));
     };
 
-    const { open, onClose } = props;
     return (
         <Dialog open={open} PaperProps={{ sx: { width: "23rem" } }}>
-            <DialogTitle>Add new result</DialogTitle>
-            <form onSubmit={(e) => writeTeamResultData(e)}>
-                <FormControlLabel
-                    className='label'
-                    label='Challenge:'
-                    labelPlacement='start'
-                    style={{ paddingBottom: "10px" }}
-                    control={
-                        <Select
-                            sx={{ minWidth: "170px", marginLeft: "25px" }}
-                            onChange={(e) => handleChallengeChange(e)}
-                        >
-                            {createChallengeItems()}
-                        </Select>
-                    }
-                />
+            <DialogTitle>Edit entry </DialogTitle>
+            <form onSubmit={(e) => updateTeamResults(e)}>
                 <FormControlLabel
                     label='Team Type: '
                     labelPlacement='start'
@@ -144,6 +139,7 @@ const NewResultDialog = (props) => {
                         <Select
                             sx={{ minWidth: "101px", marginLeft: "18px" }}
                             onChange={(e) => handleTypeChange(e)}
+                            value={teamType}
                         >
                             <MenuItem value={"Co-ed"}>Co-ed</MenuItem>
                             <MenuItem value={"Female"}>Female</MenuItem>
@@ -156,10 +152,11 @@ const NewResultDialog = (props) => {
                     labelPlacement='start'
                     style={{ paddingBottom: "10px" }}
                     onChange={(e) => handleTeamChange(e)}
+                    value={team}
                     control={
                         <TextField
                             size='small'
-                            style={{ marginLeft: "10px", width: "175px" }}
+                            style={{ marginLeft: "10px" }}
                         ></TextField>
                     }
                 />
@@ -173,10 +170,11 @@ const NewResultDialog = (props) => {
                                 <TimePicker
                                     sx={{
                                         marginLeft: "54px",
-                                        width: "175px",
                                     }}
                                     onChange={(e) => handleScoreChange(e)}
                                     views={["minutes", "seconds"]}
+                                    timeSteps={{ minutes: 1, seconds: 1 }}
+                                    value={score}
                                     format='mm:ss'
                                 />
                             </LocalizationProvider>
@@ -184,31 +182,12 @@ const NewResultDialog = (props) => {
                             <TextField
                                 size='small'
                                 type='number'
-                                style={{ marginLeft: "54px", width: "175px" }}
+                                style={{ marginLeft: "54px" }}
+                                value={score}
                             ></TextField>
                         )
                     }
                 />
-                <FormControlLabel
-                    label='Notes: '
-                    labelPlacement='start'
-                    //onChange={(e) => handleNotesChange(e)}
-                    control={
-                        <TextareaAutosize
-                            style={{
-                                fontFamily: "sans-serif",
-                                marginLeft: "3.5rem",
-                                marginTop: "10px",
-                                minRows: 10,
-                                height: "100px",
-                                maxWidth: "175px",
-                            }}
-                            maxRows={10}
-                            onChange={(e) => handleNotesChange(e)}
-                        />
-                    }
-                ></FormControlLabel>
-
                 <DialogActions>
                     <Button
                         sx={{ margin: "10px" }}
@@ -230,4 +209,4 @@ const NewResultDialog = (props) => {
         </Dialog>
     );
 };
-export default NewResultDialog;
+export default EditResultDialog;
