+++ src/Layout.jsx
@@ -33,7 +33,6 @@ const Layout = ({ children }) => {
     const open = Boolean(anchorEl);
     const [accOpen, setAccOpen] = useState(false);
     const currentChallenge = useSelector((state) => state.currentChallenge);
-    const selectedChallenge = useSelector((state) => state.selectedChallenge);
     const db = getDatabase();
     const [archivedChallgedDialogOpen, setArchivedChallengeDialogOpen] =
         useState(false);
@@ -79,11 +78,10 @@ const Layout = ({ children }) => {
         e.preventDefault();
         const challengeRef = ref(
             db,
-            "Challenges/" + selectedChallenge.replace(/ /g, "_") + "/"
+            "Challenges/" + currentChallenge.replace(/ /g, "_") + "/"
         );
         remove(challengeRef);
         setDeleteDialogOpen(false);
-        window.location.reload();
     }
 
     useEffect(() => {
@@ -216,7 +214,7 @@ const Layout = ({ children }) => {
             ></NewChallengeDialog>
             <Dialog open={deleteDialogOpen}>
                 <DialogTitle>
-                    Are you sure you want to delete {selectedChallenge}?
+                    Are you sure you want to delete {currentChallenge}?
                 </DialogTitle>
                 <DialogActions>
                     <Button
