+++ src/NotesDialog.jsx
