+++ src/firebase.js
@@ -1,5 +1,5 @@
 // Import the functions you need from the SDKs you need
-import { initializeApp, applicationDefault, cert } from "firebase/app";
+import { initializeApp } from "firebase/app";
 import { getAnalytics } from "firebase/analytics";
 import { getAuth } from "firebase/auth";
 import { getDatabase } from "firebase/database";
