+++ vite.config.js
@@ -1,0 +1,17 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  base: '/',
+  plugins: [react()],
+  build: {
+
+    /** If you set esmExternals to true, this plugins assumes that 
+      all external dependencies are ES modules */
+ 
+    commonjsOptions: {
+       esmExternals: true 
+    },
+ }
+})
